# docs/api/openapi/bundled.yaml
openapi: 3.0.4
info:
  version: 1.0.0
  title: Overlay HTTP API
  description: The overlay server API
servers:
  - url: 'http://localhost:3000'
    description: 'Optional server description, e.g. Main (production) server'
tags:
  - name: admin
    description: Admin API endpoints
  - name: non-admin
    description: Non Admin API endpointsa
paths:
  /api/v1/admin/syncAdvertisements:
    post:
      tags:
        - admin
      operationId: AdvertisementsSync
      security:
        - bearerAuth: []
      responses:
        '200':
          description: |
            Advertisement sync request successfully delegated to overlay engine.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Message:
                    type: string
                required:
                  - Message
        '500':
          $ref: '#/paths/~1api~1v1~1submit/post/responses/500'
  /api/v1/admin/startGASPSync:
    post:
      tags:
        - admin
      operationId: StartGASPSync
      security:
        - bearerAuth: []
      responses:
        '200':
          description: |
            GASP sync request successfully delegated to overlay engine.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
        '500':
          $ref: '#/paths/~1api~1v1~1submit/post/responses/500'
  /api/v1/requestForeignGASPNode:
    post:
      tags:
        - non-admin
      operationId: RequestForeignGASPNode
      parameters:
        - in: header
          name: X-BSV-Topic
          schema:
            type: string
          required: true
          description: BSV topic for the GASP node
      requestBody:
        description: Request body for the foreign GASP node endpoint
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                graphID:
                  type: string
                  description: The graph ID as an outpoint string
                txID:
                  type: string
                  description: The transaction ID
                outputIndex:
                  type: integer
                  format: uint32
                  description: The output index
              required:
                - graphID
                - txID
                - outputIndex
      responses:
        '200':
          description: |
            Foreign GASP node data.
          content:
            application/json:
              schema:
                type: object
                description: GASP node data structure
        '400':
          $ref: '#/paths/~1api~1v1~1submit/post/responses/400'
        '500':
          $ref: '#/paths/~1api~1v1~1submit/post/responses/500'
  /api/v1/requestSyncResponse:
    post:
      tags:
        - non-admin
      operationId: RequestSyncResponse
      parameters:
        - in: header
          name: X-BSV-Topic
          schema:
            type: string
          required: true
          description: BSV topic for sync response
      requestBody:
        description: Request body for the sync response endpoint
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: integer
                  description: Version of the GASP protocol
                since:
                  type: integer
                  format: uint64
                  description: Timestamp for incremental sync
              required:
                - version
                - since
      responses:
        '200':
          description: |
            Foreign GASP sync response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  utxoList:
                    type: array
                    items:
                      type: object
                      properties:
                        txid:
                          type: string
                          description: Transaction ID
                        outputIndex:
                          type: integer
                          format: uint32
                          description: Output index
                      required:
                        - txid
                        - outputIndex
                    description: List of UTXOs
                  since:
                    type: integer
                    format: uint64
                    description: Timestamp for incremental sync
                required:
                  - utxoList
                  - since
        '400':
          $ref: '#/paths/~1api~1v1~1submit/post/responses/400'
        '500':
          $ref: '#/paths/~1api~1v1~1submit/post/responses/500'
  /api/v1/submit:
    post:
      tags:
        - non-admin
      operationId: SubmitTransaction
      parameters:
        - in: header
          name: x-topics
          schema:
            type: array
            items:
              type: string
          required: true
          explode: true
          style: simple
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: object
              properties:
                Transaction:
                  type: string
                  format: binary
                  description: 'Serialized transaction data (e.g., a binary format such as Protobuf or other)'
              required:
                - Transaction
      responses:
        '200':
          description: |
            Overlay engine successfully processed the submitted transaction octet-stream with the specified topic headers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  STEAK:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        OutputsToAdmit:
                          type: array
                          items:
                            type: integer
                            format: uint32
                        CoinsToRetain:
                          type: array
                          items:
                            type: integer
                            format: uint32
                        CoinsRemoved:
                          type: array
                          items:
                            type: integer
                            format: uint32
                        AncillaryTxIDs:
                          type: array
                          items:
                            type: string
                      required:
                        - OutputsToAdmit
                        - CoinsToRetain
                        - CoinsRemoved
                        - AncillaryTxIDs
                required:
                  - STEAK
        '400':
          description: |
            The server cannot process the request due to malformed syntax or invalid input. 
            This error is typically caused by client-side issues, such as missing required parameters, 
            invalid data formats, or incorrect request structures.
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    description: Human-readable error message
                  details:
                    type: object
                    description: Optional additional error details
        '409':
          description: |
            The server did not receive a complete request within the time it was prepared to wait. 
            This error can occur if the client’s request takes too long to process or there is network latency.
          content:
            application/json:
              schema:
                $ref: '#/paths/~1api~1v1~1submit/post/responses/400/content/application~1json/schema'
        '500':
          description: |
            An unexpected condition was encountered on the server, causing the request to fail.
            This error is typically caused by internal issues such as server misconfigurations, 
            database failures, or unhandled exceptions.
          content:
            application/json:
              schema:
                $ref: '#/paths/~1api~1v1~1submit/post/responses/400/content/application~1json/schema'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer